Fixed *other* fps attribute being i64 as well and got rid of failure dependenc

+++ src/model.rs
@@ -195,7 +195,7 @@ pub struct SingleVideo {
     pub format_id: Option<String>,
     pub format_note: Option<String>,
     pub formats: Option<Vec<Format>>,
-    pub fps: Option<i64>,
+    pub fps: Option<f64>,
     pub fragment_base_url: Option<String>,
     pub fragments: Option<Vec<Fragment>>,
     pub genre: Option<String>,
